+++ src/pages/login-page/LoginPage.tsx
@@ -1,54 +1,35 @@
-// import { useState } from 'react';
-import { Link } from 'react-router-dom';
-import InputBox from '@/components/inputBox';
-import {
-  CommonLoginSection,
-  KakaoLoginButton,
-  LoginButton,
-  NaverLoginButton,
-  SocialLoginSection,
-  Wrapper
-} from './LoginPageCss';
+import { useState } from 'react';
+
+import { Wrapper } from './LoginPageCss';
+import TabButtonGroup from '@/components/tab-button';
+import OrgLogin from '@/features/login-feature/_components/org-login';
+import PersonLogin from '@/features/login-feature/_components/person-login';
 
 export default function LoginPage() {
+  const [isOrg, setIsOrg] = useState<boolean>(false);
+  const onTabChange = (tab: string) => {
+    if (tab === '기관') setIsOrg(true);
+    else setIsOrg(false);
+  };
+
   return (
     <Wrapper>
-      <i className="title">손모아 로그인</i>
-      <CommonLoginSection>
-        <form onSubmit={() => console.log('로그인폼 제출')}>
-          <div>
-            <label htmlFor="idInput">아이디</label>
-            <InputBox colortype="white" textType="text" placeholder="아이디를 입력해주세요." />
-          </div>
-          <div>
-            <label htmlFor="pwInput">비밀번호</label>
-            <InputBox colortype="white" textType="password" placeholder="비밀번호를 입력해주세요." />
-          </div>
-        </form>
-
-        <LoginButton
-          label="로그인"
-          type="blue"
-          buttonType="submit"
-          onClick={() => console.log('일반 로그인버튼 클릭')}
-        />
-        <div>
-          <Link to="/signup">회원가입</Link>
-          <Link to="/find-account">계정찾기</Link>
+      <div className="innerWrap">
+        <div className="tabWrap">
+          <TabButtonGroup
+            tabs={[{ label: '봉사자' }, { label: '기관' }]}
+            onTabChange={onTabChange}
+            width="85px"
+            height="30px"
+            borderRadius="30px"
+            fontSize="14px"
+          />
         </div>
-      </CommonLoginSection>
 
-      <SocialLoginSection>
-        <NaverLoginButton onClick={() => console.log('네이버로그인버튼클릭릭')}>
-          <img src="/assets/imgs/naver.svg" alt="네이버소셜로그인버튼" />
-          네이버로 시작하기
-        </NaverLoginButton>
-        <KakaoLoginButton onClick={() => console.log('카카오로그인버튼클릭릭')}>
-          <img src="/assets/imgs/kakao.svg" alt="카카오소셜로그인버튼" />
-          카카오로 시작하기
-        </KakaoLoginButton>
-        <p>소셜로그인은 봉사자로만 회원가입이 가능하다는 것을 알려주는 문구</p>
-      </SocialLoginSection>
+        <i className="title">손모아 로그인</i>
+
+        {isOrg ? <OrgLogin /> : <PersonLogin />}
+      </div>
     </Wrapper>
   );
 }
